Extract notification parameters to dedicated modules #3
