+++ cmk/gui/wato/_notification_parameter/registration.py
@@ -12,7 +12,6 @@ from cmk.utils.site import url_prefix
 from cmk.gui.http import request
 from cmk.gui.i18n import _
 from cmk.gui.valuespec import (
-    Alternative,
     CascadingDropdown,
     Dictionary,
     DictionaryEntry,
@@ -22,23 +21,21 @@ from cmk.gui.valuespec import (
     HTTPUrl,
     Integer,
     ListChoice,
-    Password,
     TextAreaUnicode,
     TextInput,
 )
-from cmk.gui.wato import (
-    HTTPProxyReference,
-    IndividualOrStoredPassword,
-    MigrateToIndividualOrStoredPassword,
-)
+from cmk.gui.wato import HTTPProxyReference, MigrateToIndividualOrStoredPassword
 from cmk.gui.wato._notification_parameter._helpers import get_url_prefix_specs, local_site_url
 from cmk.gui.watolib.password_store import passwordstore_choices
 
 from ._base import NotificationParameter
+from ._ilert import NotificationParameterILert
+from ._jira_issues import NotificationParameterJiraIssues
 from ._ms_teams import NotificationParameterMsTeams
 from ._opsgenie_issues import NotificationParameterOpsgenie
 from ._pushover import NotificationParameterPushover
 from ._registry import NotificationParameterRegistry
+from ._servicenow import NotificationParameterServiceNow
 from ._sms_api import NotificationParameterSMSviaIP
 from ._spectrum import NotificationParameterSpectrum
 
@@ -51,8 +48,8 @@ def register(notification_parameter_registry: NotificationParameterRegistry) -> 
     notification_parameter_registry.register(NotificationParameterPagerDuty)
     notification_parameter_registry.register(NotificationParameterSIGNL4)
     notification_parameter_registry.register(NotificationParameterASCIIMail)
-    notification_parameter_registry.register(NotificationILert)
-    notification_parameter_registry.register(NotificationParameterJIRA_ISSUES)
+    notification_parameter_registry.register(NotificationParameterILert)
+    notification_parameter_registry.register(NotificationParameterJiraIssues)
     notification_parameter_registry.register(NotificationParameterServiceNow)
     notification_parameter_registry.register(NotificationParameterOpsgenie)
     notification_parameter_registry.register(NotificationParameterSpectrum)
@@ -603,709 +600,3 @@ $LONGSERVICEOUTPUT$
         return Dictionary(
             title=_("Create notification with the following parameters"), elements=elements
         )
-
-
-class NotificationILert(NotificationParameter):
-    @property
-    def ident(self) -> str:
-        return "ilert"
-
-    @property
-    def spec(self) -> Dictionary:
-        return Dictionary(
-            title=_("Create notification with the following parameters"),
-            optional_keys=["ignore_ssl", "proxy_url"],
-            elements=[
-                (
-                    "ilert_api_key",
-                    CascadingDropdown(
-                        title=_("iLert alert source API key"),
-                        help=_("API key for iLert alert server"),
-                        choices=[
-                            (
-                                "ilert_api_key",
-                                _("API key"),
-                                TextInput(size=80, allow_empty=False),
-                            ),
-                            (
-                                "store",
-                                _("API key from password store"),
-                                DropdownChoice(sorted=True, choices=passwordstore_choices),
-                            ),
-                        ],
-                    ),
-                ),
-                (
-                    "ignore_ssl",
-                    FixedValue(
-                        value=True,
-                        title=_("Disable SSL certificate verification"),
-                        totext=_("Disable SSL certificate verification"),
-                        help=_("Ignore unverified HTTPS request warnings. Use with caution."),
-                    ),
-                ),
-                ("proxy_url", HTTPProxyReference()),
-                (
-                    "ilert_priority",
-                    DropdownChoice(
-                        sorted=True,
-                        choices=[
-                            ("HIGH", _("High (with escalation)")),
-                            ("LOW", _("Low (without escalation")),
-                        ],
-                        title=_(
-                            "Notification priority (This will override the priority configured in the alert source)"
-                        ),
-                        default_value="HIGH",
-                    ),
-                ),
-                (
-                    "ilert_summary_host",
-                    TextInput(
-                        title=_("Custom incident summary for host alerts"),
-                        default_value="$NOTIFICATIONTYPE$ Host Alert: $HOSTNAME$ is $HOSTSTATE$ - $HOSTOUTPUT$",
-                        size=64,
-                    ),
-                ),
-                (
-                    "ilert_summary_service",
-                    TextInput(
-                        title=_("Custom incident summary for service alerts"),
-                        default_value="$NOTIFICATIONTYPE$ Service Alert: $HOSTALIAS$/$SERVICEDESC$ is $SERVICESTATE$ - $SERVICEOUTPUT$",
-                        size=64,
-                    ),
-                ),
-                (
-                    "url_prefix",
-                    get_url_prefix_specs(local_site_url, default_value="automatic_https"),
-                ),
-            ],
-        )
-
-
-class NotificationParameterJIRA_ISSUES(NotificationParameter):
-    @property
-    def ident(self) -> str:
-        return "jira_issues"
-
-    @property
-    def spec(self) -> Dictionary:
-        return Dictionary(
-            title=_("Create notification with the following parameters"),
-            optional_keys=[
-                "site_customid",
-                "priority",
-                "resolution",
-                "host_summary",
-                "service_summary",
-                "ignore_ssl",
-                "timeout",
-                "label",
-            ],
-            elements=[
-                (
-                    "url",
-                    HTTPUrl(
-                        title=_("JIRA URL"),
-                        help=_("Configure the JIRA URL here."),
-                    ),
-                ),
-                (
-                    "ignore_ssl",
-                    FixedValue(
-                        value=True,
-                        title=_("Disable SSL certificate verification"),
-                        totext=_("Disable SSL certificate verification"),
-                        help=_("Ignore unverified HTTPS request warnings. Use with caution."),
-                    ),
-                ),
-                (
-                    "username",
-                    TextInput(
-                        title=_("User Name"),
-                        help=_("Configure the user name here."),
-                        size=40,
-                        allow_empty=False,
-                    ),
-                ),
-                (
-                    "password",
-                    Password(
-                        title=_("Password"),
-                        help=_(
-                            "You need to provide a valid password to be able to send notifications."
-                        ),
-                        size=40,
-                        allow_empty=False,
-                    ),
-                ),
-                (
-                    "project",
-                    TextInput(
-                        title=_("Project ID"),
-                        help=_(
-                            "The numerical JIRA project ID. If not set, it will be retrieved from a "
-                            "custom user attribute named <tt>jiraproject</tt>. "
-                            "If that is not set, the notification will fail."
-                        ),
-                        size=10,
-                    ),
-                ),
-                (
-                    "issuetype",
-                    TextInput(
-                        title=_("Issue type ID"),
-                        help=_(
-                            "The numerical JIRA issue type ID. If not set, it will be retrieved from a "
-                            "custom user attribute named <tt>jiraissuetype</tt>. "
-                            "If that is not set, the notification will fail."
-                        ),
-                        size=10,
-                    ),
-                ),
-                (
-                    "host_customid",
-                    TextInput(
-                        title=_("Host custom field ID"),
-                        help=_("The numerical JIRA custom field ID for host problems."),
-                        size=10,
-                    ),
-                ),
-                (
-                    "service_customid",
-                    TextInput(
-                        title=_("Service custom field ID"),
-                        help=_("The numerical JIRA custom field ID for service problems."),
-                        size=10,
-                    ),
-                ),
-                (
-                    "site_customid",
-                    TextInput(
-                        title=_("Site custom field ID"),
-                        help=_(
-                            "The numerical ID of the JIRA custom field for sites. "
-                            "Please use this option if you have multiple sites in a "
-                            "distributed setup which send their notifications "
-                            "to the same JIRA instance."
-                        ),
-                        size=10,
-                    ),
-                ),
-                (
-                    "monitoring",
-                    HTTPUrl(
-                        title=_("Monitoring URL"),
-                        help=_(
-                            "Configure the base URL for the Monitoring Web-GUI here. Include the site name. "
-                            "Used for link to check_mk out of jira."
-                        ),
-                    ),
-                ),
-                (
-                    "priority",
-                    TextInput(
-                        title=_("Priority ID"),
-                        help=_(
-                            "The numerical JIRA priority ID. If not set, it will be retrieved from a "
-                            "custom user attribute named <tt>jirapriority</tt>. "
-                            "If that is not set, the standard priority will be used."
-                        ),
-                        size=10,
-                    ),
-                ),
-                (
-                    "host_summary",
-                    TextInput(
-                        title=_("Summary for host notifications"),
-                        help=_(
-                            "Here you are allowed to use all macros that are defined in the "
-                            "notification context."
-                        ),
-                        default_value="Check_MK: $HOSTNAME$ - $HOSTSHORTSTATE$",
-                        size=64,
-                    ),
-                ),
-                (
-                    "service_summary",
-                    TextInput(
-                        title=_("Summary for service notifications"),
-                        help=_(
-                            "Here you are allowed to use all macros that are defined in the "
-                            "notification context."
-                        ),
-                        default_value="Check_MK: $HOSTNAME$/$SERVICEDESC$ $SERVICESHORTSTATE$",
-                        size=64,
-                    ),
-                ),
-                (
-                    "label",
-                    TextInput(
-                        title=_("Label"),
-                        help=_(
-                            "Here you can set a custom label for new issues. "
-                            "If not set, 'monitoring' will be used."
-                        ),
-                        size=16,
-                    ),
-                ),
-                (
-                    "resolution",
-                    TextInput(
-                        title=_("Activate resolution with following resolution transition ID"),
-                        help=_(
-                            "The numerical JIRA resolution transition ID. "
-                            "11 - 'To Do', 21 - 'In Progress', 31 - 'Done'"
-                        ),
-                        size=3,
-                    ),
-                ),
-                (
-                    "timeout",
-                    TextInput(
-                        title=_("Set optional timeout for connections to JIRA"),
-                        help=_("Here you can configure timeout settings."),
-                        default_value="10",
-                    ),
-                ),
-            ],
-        )
-
-
-class NotificationParameterServiceNow(NotificationParameter):
-    @property
-    def ident(self) -> str:
-        return "servicenow"
-
-    @property
-    def spec(self) -> Dictionary:
-        return Dictionary(
-            title=_("Create notification with the following parameters"),
-            required_keys=["url", "username", "password", "mgmt_type"],
-            elements=[
-                (
-                    "url",
-                    HTTPUrl(
-                        title=_("ServiceNow URL"),
-                        help=_(
-                            "Configure your ServiceNow URL here (eg. https://myservicenow.com)."
-                        ),
-                        allow_empty=False,
-                    ),
-                ),
-                ("proxy_url", HTTPProxyReference()),
-                (
-                    "username",
-                    TextInput(
-                        title=_("Username"),
-                        help=_(
-                            "The user, used for login, has to have at least the "
-                            "role 'itil' in ServiceNow."
-                        ),
-                        size=40,
-                        allow_empty=False,
-                    ),
-                ),
-                (
-                    "password",
-                    IndividualOrStoredPassword(
-                        title=_("Password of the user"),
-                        allow_empty=False,
-                    ),
-                ),
-                (
-                    "use_site_id",
-                    Alternative(
-                        title=_("Use site ID prefix"),
-                        help=_(
-                            "Please use this option if you have multiple "
-                            "sites in a distributed setup which send their "
-                            "notifications to the same ServiceNow instance. "
-                            "The site ID will be used as prefix for the "
-                            "problem ID on incident creation."
-                        ),
-                        elements=[
-                            FixedValue(value=False, title=_("Deactivated"), totext=""),
-                            FixedValue(value=True, title=_("Use site ID"), totext=""),
-                        ],
-                        default_value=False,
-                    ),
-                ),
-                (
-                    "timeout",
-                    TextInput(
-                        title=_("Set optional timeout for connections to ServiceNow"),
-                        help=_("Here you can configure timeout settings in seconds."),
-                        default_value="10",
-                        size=3,
-                    ),
-                ),
-                (
-                    "mgmt_type",
-                    CascadingDropdown(
-                        title=_("Management type"),
-                        help=_(
-                            "With ServiceNow you can create different "
-                            "types of management issues, currently "
-                            "supported are indicents and cases."
-                        ),
-                        choices=[
-                            ("incident", _("Incident"), self._incident_vs()),
-                            ("case", _("Case"), self._case_vs()),
-                        ],
-                    ),
-                ),
-            ],
-        )
-
-    def _incident_vs(self) -> Dictionary:
-        return Dictionary(
-            title=_("Incident"),
-            required_keys=["caller"],
-            elements=[
-                (
-                    "caller",
-                    TextInput(
-                        title=_("Caller ID"),
-                        help=_(
-                            "Caller is the user on behalf of whom the incident is being reported "
-                            "within ServiceNow. Please enter the name of the caller here. "
-                            "It is recommended to use the same user as used for login. "
-                            "Otherwise, your ACL rules in ServiceNow must be "
-                            "adjusted, so that the user who is used for login "
-                            "can create/edit/resolve incidents on behalf of the "
-                            "caller. Please have a look at ServiceNow "
-                            "documentation for details."
-                        ),
-                    ),
-                ),
-                ("host_short_desc", self._host_short_desc(_("incidents"))),
-                ("svc_short_desc", self._svc_short_desc(_("incidents"))),
-                (
-                    "host_desc",
-                    self._host_desc(
-                        title=_("Description for host incidents"),
-                        help_text=_(
-                            "Text that should be set in field <tt>Description</tt> "
-                            "for host notifications."
-                        ),
-                    ),
-                ),
-                (
-                    "svc_desc",
-                    self._svc_desc(
-                        title=_("Description for service incidents"),
-                        help_text=_(
-                            "Text that should be set in field <tt>Description</tt> "
-                            "for service notifications."
-                        ),
-                    ),
-                ),
-                (
-                    "urgency",
-                    DropdownChoice(
-                        title=_("Urgency"),
-                        help=_(
-                            'See <a href="https://docs.servicenow.com/bundle/'
-                            "helsinki-it-service-management/page/product/incident-management/"
-                            'reference/r_PrioritizationOfIncidents.html" target="_blank">'
-                            "ServiceNow Incident</a> for more information."
-                        ),
-                        choices=[
-                            ("low", _("Low")),
-                            ("medium", _("Medium")),
-                            ("high", _("High")),
-                        ],
-                        default_value="low",
-                    ),
-                ),
-                (
-                    "impact",
-                    DropdownChoice(
-                        title=_("Impact"),
-                        help=_(
-                            'See <a href="https://docs.servicenow.com/bundle/'
-                            "helsinki-it-service-management/page/product/incident-management/"
-                            'reference/r_PrioritizationOfIncidents.html" target="_blank">'
-                            "ServiceNow Incident</a> for more information."
-                        ),
-                        choices=[
-                            ("low", _("Low")),
-                            ("medium", _("Medium")),
-                            ("high", _("High")),
-                        ],
-                        default_value="low",
-                    ),
-                ),
-                (
-                    "ack_state",
-                    Dictionary(
-                        title=_("Settings for incident state in case of acknowledgement"),
-                        help=_(
-                            "Here you can define the state of the incident in case of an "
-                            "acknowledgement of the affected host or service problem."
-                        ),
-                        elements=[
-                            (
-                                "start",
-                                Alternative(
-                                    title=_("State of incident if acknowledgement is set"),
-                                    help=_(
-                                        "Here you can define the state of the incident in case of an "
-                                        "acknowledgement of the host or service problem."
-                                    ),
-                                    elements=[
-                                        DropdownChoice(
-                                            title=_(
-                                                "State of incident if acknowledgement is set (predefined)"
-                                            ),
-                                            help=_(
-                                                "Please note that the mapping to the numeric "
-                                                "ServiceNow state may be changed at your system "
-                                                "and can differ from our definitions. In this case "
-                                                "use the option below."
-                                            ),
-                                            choices=self._get_state_choices("incident"),
-                                            default_value="none",
-                                        ),
-                                        Integer(
-                                            title=_(
-                                                "State of incident if acknowledgement is set (as integer)"
-                                            ),
-                                            minvalue=0,
-                                        ),
-                                    ],
-                                ),
-                            ),
-                        ],
-                    ),
-                ),
-                ("recovery_state", self._recovery_state_vs(_("incident"))),
-                (
-                    "dt_state",
-                    Dictionary(
-                        title=_("Settings for incident state in case of downtime"),
-                        help=_(
-                            "Here you can define the state of the incident in case of a "
-                            "downtime of the affected host or service."
-                        ),
-                        elements=[
-                            (
-                                "start",
-                                Alternative(
-                                    title=_("State of incident if downtime is set"),
-                                    elements=[
-                                        DropdownChoice(
-                                            title=_(
-                                                "State of incident if downtime is set (predefined)"
-                                            ),
-                                            help=_(
-                                                "Please note that the mapping to the numeric "
-                                                "ServiceNow state may be changed at your system "
-                                                "and can differ from our definitions. In this case "
-                                                "use the option below."
-                                            ),
-                                            choices=self._get_state_choices("incident"),
-                                            default_value="none",
-                                        ),
-                                        Integer(
-                                            title=_(
-                                                "State of incident if downtime is set (as integer)"
-                                            ),
-                                            minvalue=0,
-                                        ),
-                                    ],
-                                ),
-                            ),
-                            (
-                                "end",
-                                Alternative(
-                                    title=_("State of incident if downtime expires"),
-                                    help=_(
-                                        "Here you can define the state of the incident in case of an "
-                                        "ending acknowledgement of the host or service problem."
-                                    ),
-                                    elements=[
-                                        DropdownChoice(
-                                            title=_(
-                                                "State of incident if downtime expires (predefined)"
-                                            ),
-                                            help=_(
-                                                "Please note that the mapping to the numeric "
-                                                "ServiceNow state may be changed at your system "
-                                                "and can differ from our definitions. In this case "
-                                                "use the option below."
-                                            ),
-                                            choices=self._get_state_choices("incident"),
-                                            default_value="none",
-                                        ),
-                                        Integer(
-                                            title=_(
-                                                "State of incident if downtime expires (as integer)"
-                                            ),
-                                            minvalue=0,
-                                        ),
-                                    ],
-                                ),
-                            ),
-                        ],
-                    ),
-                ),
-            ],
-        )
-
-    def _case_vs(self) -> Dictionary:
-        return Dictionary(
-            title=_("Case"),
-            elements=[
-                ("host_short_desc", self._host_short_desc(_("cases"))),
-                ("svc_short_desc", self._svc_short_desc(_("cases"))),
-                (
-                    "host_desc",
-                    self._host_desc(
-                        title=_("Resolution notes for host cases"),
-                        help_text=_(
-                            "Text that should be set in field <tt>Resolution notes</tt> "
-                            "on recovery service notifications."
-                        ),
-                    ),
-                ),
-                (
-                    "svc_desc",
-                    self._svc_desc(
-                        title=_("Resolution notes for service cases"),
-                        help_text=_(
-                            "Text that should be set in field <tt>Resolution notes</tt> "
-                            "on recovery service notifications."
-                        ),
-                    ),
-                ),
-                (
-                    "priority",
-                    DropdownChoice(
-                        title=_("Priority"),
-                        help=_(
-                            "Here you can define with which priority the case should be created."
-                        ),
-                        choices=[
-                            ("low", _("Low")),
-                            ("moderate", _("Moderate")),
-                            ("high", _("High")),
-                            ("critical", _("Critical")),
-                        ],
-                        default_value="low",
-                    ),
-                ),
-                ("recovery_state", self._recovery_state_vs(_("case"))),
-            ],
-        )
-
-    def _host_desc(self, title: str, help_text: str) -> TextAreaUnicode:
-        return TextAreaUnicode(
-            title=title,
-            help=help_text,
-            rows=7,
-            cols=58,
-            monospaced=True,
-            default_value="""Host: $HOSTNAME$
-Event:    $EVENT_TXT$
-Output:   $HOSTOUTPUT$
-Perfdata: $HOSTPERFDATA$
-$LONGHOSTOUTPUT$
-""",
-        )
-
-    def _svc_desc(self, title: str, help_text: str) -> TextAreaUnicode:
-        return TextAreaUnicode(
-            title=title,
-            help=help_text,
-            rows=11,
-            cols=58,
-            monospaced=True,
-            default_value="""Host: $HOSTNAME$
-Service:  $SERVICEDESC$
-Event:    $EVENT_TXT$
-Output:   $SERVICEOUTPUT$
-Perfdata: $SERVICEPERFDATA$
-$LONGSERVICEOUTPUT$
-""",
-        )
-
-    def _host_short_desc(self, issue_type: str) -> TextInput:
-        return TextInput(
-            title=_("Short description for host %s") % issue_type,
-            help=_(
-                "Text that should be set in field <tt>Short description</tt> "
-                "for host notifications."
-            ),
-            default_value="Check_MK: $HOSTNAME$ - $HOSTSHORTSTATE$",
-            size=64,
-        )
-
-    def _svc_short_desc(self, issue_type: str) -> TextInput:
-        return TextInput(
-            title=_("Short description for service %s") % issue_type,
-            help=_(
-                "Text that should be set in field <tt>Short description</tt> "
-                "for service notifications."
-            ),
-            default_value="Check_MK: $HOSTNAME$/$SERVICEDESC$ $SERVICESHORTSTATE$",
-            size=68,
-        )
-
-    def _recovery_state_vs(self, issue_type: str) -> Dictionary:
-        return Dictionary(
-            title=_("Settings for %s state in case of recovery") % issue_type,
-            help=_(
-                "Here you can define the state of the %s in case of a recovery "
-                "of the affected host or service problem."
-            )
-            % issue_type,
-            elements=[
-                (
-                    "start",
-                    Alternative(
-                        title=_("State of %s if recovery is set") % issue_type,
-                        elements=[
-                            DropdownChoice(
-                                title=_("State of case if recovery is set (predefined)"),
-                                help=_(
-                                    "Please note that the mapping to the numeric "
-                                    "ServiceNow state may be changed at your system "
-                                    "and can differ from our definitions. In this case "
-                                    "use the option below."
-                                ),
-                                choices=self._get_state_choices(issue_type),
-                                default_value="none",
-                            ),
-                            Integer(
-                                title=_("State of %s if recovery is set (as integer)") % issue_type,
-                                minvalue=0,
-                            ),
-                        ],
-                    ),
-                ),
-            ],
-        )
-
-    def _get_state_choices(self, issue_type: str) -> list[tuple[str, str]]:
-        if issue_type == "incident":
-            return [
-                ("none", _("Don't change state")),
-                ("new", _("New")),
-                ("progress", _("In Progress")),
-                ("hold", _("On Hold")),
-                ("resolved", _("Resolved")),
-                ("closed", _("Closed")),
-                ("canceled", _("Canceled")),
-            ]
-
-        # Cases
-        return [
-            ("none", _("Don't change state")),
-            ("new", _("New")),
-            ("closed", _("Closed")),
-            ("resolved", _("Resolved")),
-            ("open", _("Open")),
-            ("awaiting_info", _("Awaiting info")),
-        ]
